feat: add migration to enable tenant level toggle
